requirements.txt
annotated-types==0.7.0
anyio==4.11.0
black==25.9.0
click==8.3.0
fastapi==0.118.0
h11==0.16.0
httpx==0.28.1
idna==3.10
isort==6.1.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pydantic==2.11.10
pydantic_core==2.33.2
pytokens==0.1.10
python-dateutil==2.9.0
redis==5.2.1
ruff==0.13.3
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
eth-abi==5.1.0
eth-utils==4.1.1
eth-account==0.11.3
motor==3.3.2
pymongo==4.6.1
